ref(ui): Refactor `<FrameRegisters>`
* Convert to ts
* Remove `grid-emotion`
